moved HTML parser into '_main_/utils/html_parser/'
